Moved more code to separate files to get more visibility. Move Find target fom Ally to Player. Moved enemy and ally list to functions arguments instead of using the global variables. Moved quest  function from npc to npc interaction function. Fixed a bit with items.
