+++ branches.py
@@ -1,9 +1,9 @@
-from utilities import *
+from utilities import clear_console
 from room import Room
 
 # -- Branches --
 #Prints and lets the player choose where to go to next
-def road_choice(player, room_1: Room, room_2: Room, room_3: Room = ""):
+def road_choice(player, enemy_list, ally_list, room_1: Room, room_2: Room, room_3: Room = ""):
     i = 0
     while i < 1:
         print("Where do you want to go next?")
@@ -19,9 +19,9 @@ def road_choice(player, room_1: Room, room_2: Room, room_3: Room = ""):
         clear_console()
         #Entering room depending on choice
         if choice == 1:
-            room_1.enter_room(player)
+            room_1.enter_room(player, enemy_list, ally_list)
         elif choice == 2:
-            room_2.enter_room(player)
+            room_2.enter_room(player, enemy_list, ally_list)
         else:
-            room_3.enter_room(player)
+            room_3.enter_room(player, enemy_list, ally_list)
         i+=1
