+++ player.py
@@ -1,7 +1,6 @@
-from entity import *
-from item import *
-
-# --- Player class ---
+from utilities import row_devider
+from entites.ally import Ally
+from item import Equipable, Usable
 
 class Player(Ally):
     symbol = '''     _
@@ -75,3 +74,54 @@ class Player(Ally):
     #What is this used for???
     def player_stats(self, item):
         print(f"Statistics: \nDamage: {self.damage} + {item.bonus} \nHealth: {self.hp}")
+
+    def Find_target(self, enemy_list, ally_list):
+        #print(f"what do you want to do using {self}\n")
+        #self.print_self()
+        while True:
+            row_devider()
+            print("What will you do?")
+            print("1. Attack")
+            print("2. Use Item")
+            print("3. Check Stats and inventory")
+            action = int(input("Choose an action (1, 2 or 3):\n "))
+            row_devider()
+
+            if action == 1:
+                index = 0
+                if len(enemy_list) > 1:
+                    print("What enemy do you want to attack\n")
+                    i = 1
+                    for enemy in enemy_list:
+                        print(f"{i}: {enemy.name}: {enemy.hp}/{enemy.max_hp}")
+                        i+= 1
+                    index = int(input()) - 1
+                target = enemy_list[index]
+
+                killed = self.attack(target)
+                if killed:
+                    enemy_list.remove(target)
+                    print(f"You killed a {target.name}!")
+
+                break
+
+            elif action == 2:
+                used = self.use_items()
+
+                if not used:
+                    continue
+
+            elif action == 3:
+                for ally in ally_list:
+                    ally.print_self()
+
+                print("\nInventory:")
+                #Skriva ut inventory !!!!!!!!!!!!
+                continue
+            else:
+                print("Invalid action. You hesitate...")
+                continue
+
+            #If all enemies in room is deafeted
+            if len(enemy_list) < 1:
+                break
